refactor: centralize command execution
